third_party/blink/web_tests/fast/dom/HTMLLinkElement/link-preload-validity.html

<!DOCTYPE html>
<script>
    if (window.testRunner) {
        testRunner.dumpAsText();
        testRunner.dumpChildFrames();
    }
    if (window.internals)
        internals.settings.setLogPreload(true);
</script>
<link rel="preload" as="fetch" href="http://prefetch.wut.com.test/1.gif">
<link rel="preload" as="image" href="ftp:???://prefetch.wut.com.test/2.gif">
<link rel="preload" as="image" href="">
<link rel="preload" as="image">
<link rel="preload" as="image" href="http://prefetch.wut.com.test/3.gif">
This test verifies that a preload request is sent and logged.
